perf(sCMOS): vectorize gaussian_filter_stack; fix float64 leak in demosaic paths
gaussian_filter_stack looped skimage.filters.gaussian per frame. skimage's
gaussian is a thin wrapper around scipy.ndimage.gaussian_filter (already
separable, mode='nearest', truncate=4.0), so the loop + per-call overhead
was the actual cost, not the convolution algorithm. Replaced with a single
scipy.ndimage.gaussian_filter call per stack (sigma=(0, sigma, sigma) leaves
the frame axis untouched) -- bit-identical output, ~15-18x faster at
realistic scale (14x14 puncta stacks), 1.6-2.4x at larger sigma/ROI.

Also fixed a real float64 leak on the DEMOSAIC/STANDARD_IG fitting
strategies: bayer_demosaic_stack's RGB_image = np.zeros([...]) and
multicolour.py's _bayer_destacker's destacked_image = np.zeros([...]) both
lacked dtype=, silently upcasting to float64 despite float32 input.
STANDARD/ELLIPTICAL paths were already float32 throughout.

Tests: unit_tests/test_scmos_functions.py -- added exact-match-vs-reference
and frame-independence regression tests for the vectorized smoothing.
sCMOSFunctions.py at 100% coverage; full multicolour/Nile Red/AnalysisPipeline/
SR_Functions/demosaic-strategy/photoelectron-extraction suites re-run, no
regressions (278 passed).

See claude/LOG.md "August 17, 2026" sessions for full benchmark numbers.

+++ src/simulation/multicolour.py
@@ -676,7 +676,8 @@ class MultiC_Sim_Funcs_Refactored:
                        each frame contains data from a single colour channel
         """
         destacked_image = np.zeros(
-            [RGB_image.shape[0] * 3, RGB_image.shape[1], RGB_image.shape[2]]
+            [RGB_image.shape[0] * 3, RGB_image.shape[1], RGB_image.shape[2]],
+            dtype=RGB_image.dtype,
         )
         index = 0
         for i in range(RGB_image.shape[0]):
